requirements.txt
google-genai==0.3.0
supabase==2.3.4
python-dotenv==1.0.0
requests==2.31.0
boto3==1.34.0
edge-tts==6.1.9
mutagen==1.47.0
google-api-python-client==2.118.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0